Refactor dashboard endpoints into one file.
