+++ src/frontend-admin/src/components/reusables/PieChart/CustomLabel/CustomLabel.tsx
@@ -27,7 +27,7 @@ export const CustomLabel: React.FunctionComponent<ICustomLabelProps> = props => 
 
     var radius
     var textColor
-    console.log(data[index].value)
+    // console.log(data[index].value)
     if (data[index].value == 0) {
         radius = innerRadius + (outerRadius - innerRadius) * 1.15
         textColor = '#f4f4f4'
